refactor: move services from `utils/` to `services/` (#171)
