+++ test/server/api/mocks/transition-page.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-classes-per-file, class-methods-use-this */
 import { injectable } from 'inversify'
-import { CookieReducer } from '../../../../src/server/util/transition-page'
+import { CookieReducer } from '../../../../src/server/services/transition-page'
 
 @injectable()
 export class CookieArrayReducerMockVisited implements CookieReducer {
